Don't put double quotes around string params
This is a deviation from standard JSON RPC that we allow because MQTT
payloads are 'self-delimiting' and we don't need to know their size.
So, rather than having a return value of:

"OK"

we permit simply:

OK

with the understanding that *everything* returned by the API is a
string. The same applies to numbers, e.g., we don't want to see:

"24"

but instead:

24

However, strings seem to have been the only 'problematic' type in this
regard in the jrgen codebase.